Migrate `ExcludeInternalDoclet` test
